Adds get offers by city function
